Move GitHub source validation to separate file
