+++ spec/concerns/webhook_validations_spec.rb
@@ -25,12 +25,4 @@ describe WebhookValidations do
     klass = WebhookValidationsTester.new("127.0.0.1")
     expect(klass).to_not be_valid_incoming_webhook_address
   end
-
-  context "verifies IPs" do
-    it "returns production" do
-      expect(WebhookValidations::Validator.new("127.0.0.1")).to_not be_valid
-      expect(WebhookValidations::Validator.new("192.30.252.41")).to be_valid
-      expect(WebhookValidations::Validator.new("192.30.252.46")).to be_valid
-    end
-  end
 end
